# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ticket Booking System API Documentation",
        "description": "API documentation for the Ticket Booking System backend.",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Ticket Booking System API Server"
        }
    ],
    "paths": {
        "/admin/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get all users (Admin only)",
                "operationId": "afab67256b9405de17d3e13f2eae9c4e",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminUserList"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Create a new user (Admin only)",
                "operationId": "096428f5a3b73c7535b724cfc91b53b5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "role"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "New Admin"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "newadmin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "user",
                                            "admin"
                                        ],
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminUser"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/users/{user}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update an existing user (Admin only)",
                "operationId": "53389c40d1fb113820fa9ffbcfc544b4",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID of the user to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Name"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "updated@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "new_password"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "user",
                                            "admin"
                                        ],
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminUser"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete a user (Admin only)",
                "operationId": "a11154970ac963d09992c64fddc6b6d0",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID of the user to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User deleted successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/bookings": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get all bookings (Admin only)",
                "operationId": "67a5de392c293c3e74d1e5403af20b82",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminBookingList"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/bookings/{booking}/status": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update booking status (Admin only)",
                "operationId": "f8c648a237f8e7b1a8fd1dc96f2b2be4",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "description": "ID of the booking to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "confirmed",
                                            "cancelled"
                                        ],
                                        "example": "confirmed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Booking"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Booking not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/bookings/{booking}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete a booking (Admin only)",
                "operationId": "c823fbbba4a338aa7794e6c4e29343e2",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "description": "ID of the booking to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Booking deleted successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Booking not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AdminUser": {
                "title": "Admin User",
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ValidationError"
                    },
                    {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "format": "int64",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Admin User"
                            },
                            "email": {
                                "type": "string",
                                "format": "email",
                                "example": "admin@example.com"
                            },
                            "role": {
                                "type": "string",
                                "example": "admin"
                            },
                            "created_at": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2025-01-01T00:00:00.000000Z"
                            },
                            "updated_at": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2025-01-01T00:00:00.000000Z"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "AdminUserList": {
                "title": "Admin User List",
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/AdminUser"
                },
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ValidationError"
                    }
                ]
            },
            "AdminBookingList": {
                "title": "Admin Booking List",
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/Booking"
                },
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ValidationError"
                    }
                ]
            },
            "Booking": {
                "title": "Booking",
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ValidationError"
                    },
                    {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "format": "int64",
                                "example": 1
                            },
                            "user_id": {
                                "type": "integer",
                                "format": "int64",
                                "example": 1
                            },
                            "from_station": {
                                "type": "string",
                                "example": "Station A"
                            },
                            "to_station": {
                                "type": "string",
                                "example": "Station B"
                            },
                            "journey_date": {
                                "type": "string",
                                "format": "date",
                                "example": "2025-12-25"
                            },
                            "seat_type": {
                                "type": "string",
                                "example": "Economy"
                            },
                            "number_of_seats": {
                                "type": "integer",
                                "example": 2
                            },
                            "total_fare": {
                                "type": "number",
                                "format": "float",
                                "example": 100
                            },
                            "status": {
                                "type": "string",
                                "example": "pending"
                            },
                            "created_at": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2025-01-01T00:00:00.000000Z"
                            },
                            "updated_at": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2025-01-01T00:00:00.000000Z"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "ValidationError": {
                "title": "Validation Error",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}